Introduce workload channel struct
